style(web-demo): move customize into it's own folder
