+++ packages/pixelmanipulator-web-demo/src/components/customize/CustomColorAlpha.vue
