+++ packages/pixelmanipulator-web-demo/src/components/customize/CustomizeColor.vue
