+++ packages/pixelmanipulator-web-demo/src/components/customize/CustomizeName.vue
