+++ packages/pixelmanipulator-web-demo/src/components/customize/ElementCustomize.vue
